ui/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in --output-file requirements.txt
annotated-types==0.7.0
    # via pydantic
anthropic==0.42.0
    # via claudette
anyio==4.8.0
    # via
    #   anthropic
    #   httpx
    #   starlette
    #   watchfiles
apsw==3.47.2.0
    # via apswutils
apswutils==0.0.2
    # via fastlite
beautifulsoup4==4.12.3
    # via
    #   python-fasthtml
    #   toolslm
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
claudette==0.1.1
    # via -r requirements.in
click==8.1.8
    # via uvicorn
distro==1.9.0
    # via anthropic
fastcore==1.7.28
    # via
    #   apswutils
    #   claudette
    #   fastlite
    #   llms-txt
    #   msglm
    #   python-fasthtml
    #   toolslm
fastlite==0.1.1
    # via python-fasthtml
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
html2text==2024.2.26
    # via toolslm
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   anthropic
    #   llms-txt
    #   python-fasthtml
    #   toolslm
idna==3.10
    # via
    #   anyio
    #   httpx
itsdangerous==2.2.0
    # via python-fasthtml
jiter==0.8.2
    # via anthropic
llms-txt==0.0.4
    # via toolslm
msglm==0.0.4
    # via claudette
oauthlib==3.2.2
    # via python-fasthtml
packaging==24.2
    # via fastcore
pydantic==2.10.5
    # via anthropic
pydantic-core==2.27.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via python-fasthtml
python-dotenv==1.0.1
    # via uvicorn
python-fasthtml==0.12.0
    # via -r requirements.in
python-multipart==0.0.20
    # via python-fasthtml
pyyaml==6.0.2
    # via uvicorn
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
soupsieve==2.6
    # via beautifulsoup4
starlette==0.45.2
    # via python-fasthtml
toolslm==0.1.0
    # via claudette
typing-extensions==4.12.2
    # via
    #   anthropic
    #   anyio
    #   pydantic
    #   pydantic-core
uvicorn==0.34.0
    # via python-fasthtml
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.4
    # via uvicorn
websockets==14.1
    # via uvicorn
